added code to the calendarEvents

+++ app/src/main/java/team04/goalsmasher/controller/CalendarEventSchedule.java
@@ -26,7 +26,7 @@ public class CalendarEventSchedule extends AppCompatActivity {
         calendarView.setOnDateChangeListener(new CalendarView.OnDateChangeListener() {
             @Override
             public void onSelectedDayChange(@NonNull CalendarView view, int year, int month, int dayOfMonth) {
-                String date = (year) + "/" + month + 1 + "/" + dayOfMonth;
+                String date = (year) + "/" + (int)(month + 1) + "/" + dayOfMonth;
                 calendarGoals.setText(date);
             }
         });
